Handle floating point drift in larva timing
After reaching an integer frame that is past the floating point partial
frame where the larva was actually spawned, the new "goal post" frame is
set based on the current integer frame. This causes drift over time.
Updating based on the last "goal post" frame will account for that drift
so that it never goes more than one frame. Also, checking if the current
frame is past that "goal post" will introduce drift if "goal post" is
ever an integer, it will just skip a whole frame, so the check should be
"<=", not "<".

Signed-off-by: dusoleil <howcansocksbereal@gmail.com>