Init larva timer in game tick
Initializing the larva timer in the constructor bases it on game frame
0, which causes a bug for every hatch after the first.
We can check if it is uninitialized in the game tick and set it based on
the current frame.
We also need to give a single larva if there aren't any because hatches
spawn with one larva (except for the first which spawns with 3).
Initializing the larva count in the constructor also causes a weird bug
that I couldn't figure out, so I'm also initializing it in the game
tick.

Signed-off-by: dusoleil <howcansocksbereal@gmail.com>